# Automatic SSE color converters based on Vector API