--- a/src/ImageSharp/Common/Helpers/SimdUtils.cs
+++ b/src/ImageSharp/Common/Helpers/SimdUtils.cs
@@ -33,18 +33,6 @@ namespace SixLabors.ImageSharp
         public static bool HasVector4 { get; } =
             Vector.IsHardwareAccelerated && Vector<float>.Count == 4;
 
-        public static bool HasAvx2
-        {
-            get
-            {
-#if SUPPORTS_RUNTIME_INTRINSICS
-                return Avx2.IsSupported;
-#else
-                return false;
-#endif
-            }
-        }
-
         /// <summary>
         /// Transform all scalars in 'v' in a way that converting them to <see cref="int"/> would have rounding semantics.
         /// </summary>