--- a/src/ImageSharp/Formats/Jpeg/Components/Decoder/ColorConverters/JpegColorConverter.FromCmykVector.cs
+++ b/src/ImageSharp/Formats/Jpeg/Components/Decoder/ColorConverters/JpegColorConverter.FromCmykVector.cs
@@ -9,38 +9,42 @@ namespace SixLabors.ImageSharp.Formats.Jpeg.Components.Decoder.ColorConverters
 {
     internal abstract partial class JpegColorConverterBase
     {
-        internal sealed class FromRgbVector8 : VectorizedJpegColorConverter
+        internal sealed class FromCmykVector : VectorizedJpegColorConverter
         {
-            public FromRgbVector8(int precision)
-                : base(JpegColorSpace.RGB, precision)
+            public FromCmykVector(int precision)
+                : base(JpegColorSpace.Cmyk, precision)
             {
             }
 
             protected override void ConvertCoreVectorizedInplace(in ComponentValues values)
             {
-                ref Vector<float> rBase =
+                ref Vector<float> cBase =
                     ref Unsafe.As<float, Vector<float>>(ref MemoryMarshal.GetReference(values.Component0));
-                ref Vector<float> gBase =
+                ref Vector<float> mBase =
                     ref Unsafe.As<float, Vector<float>>(ref MemoryMarshal.GetReference(values.Component1));
-                ref Vector<float> bBase =
+                ref Vector<float> yBase =
                     ref Unsafe.As<float, Vector<float>>(ref MemoryMarshal.GetReference(values.Component2));
+                ref Vector<float> kBase =
+                    ref Unsafe.As<float, Vector<float>>(ref MemoryMarshal.GetReference(values.Component3));
 
                 var scale = new Vector<float>(1 / this.MaximumValue);
 
-                nint n = values.Component0.Length / 8;
+                nint n = values.Component0.Length / Vector<float>.Count;
                 for (nint i = 0; i < n; i++)
                 {
-                    ref Vector<float> r = ref Unsafe.Add(ref rBase, i);
-                    ref Vector<float> g = ref Unsafe.Add(ref gBase, i);
-                    ref Vector<float> b = ref Unsafe.Add(ref bBase, i);
-                    r *= scale;
-                    g *= scale;
-                    b *= scale;
+                    ref Vector<float> c = ref Unsafe.Add(ref cBase, i);
+                    ref Vector<float> m = ref Unsafe.Add(ref mBase, i);
+                    ref Vector<float> y = ref Unsafe.Add(ref yBase, i);
+                    Vector<float> k = Unsafe.Add(ref kBase, i) * scale;
+
+                    c = c * k * scale;
+                    m = m * k * scale;
+                    y = y * k * scale;
                 }
             }
 
             protected override void ConvertCoreInplace(in ComponentValues values) =>
-                FromRgbScalar.ConvertCoreInplace(values, this.MaximumValue);
+                FromCmykScalar.ConvertCoreInplace(values, this.MaximumValue);
         }
     }
 }