--- a/src/ImageSharp/Formats/Jpeg/Components/Decoder/ColorConverters/JpegColorConverter.FromCmykVector.cs
+++ b/src/ImageSharp/Formats/Jpeg/Components/Decoder/ColorConverters/JpegColorConverter.FromCmykVector.cs
@@ -27,7 +27,7 @@ namespace SixLabors.ImageSharp.Formats.Jpeg.Components.Decoder.ColorConverters
                 ref Vector<float> kBase =
                     ref Unsafe.As<float, Vector<float>>(ref MemoryMarshal.GetReference(values.Component3));
 
-                var scale = new Vector<float>(1 / this.MaximumValue);
+                var scale = new Vector<float>(1 / (this.MaximumValue * this.MaximumValue));
 
                 nint n = values.Component0.Length / Vector<float>.Count;
                 for (nint i = 0; i < n; i++)
@@ -35,11 +35,12 @@ namespace SixLabors.ImageSharp.Formats.Jpeg.Components.Decoder.ColorConverters
                     ref Vector<float> c = ref Unsafe.Add(ref cBase, i);
                     ref Vector<float> m = ref Unsafe.Add(ref mBase, i);
                     ref Vector<float> y = ref Unsafe.Add(ref yBase, i);
-                    Vector<float> k = Unsafe.Add(ref kBase, i) * scale;
+                    Vector<float> k = Unsafe.Add(ref kBase, i);
 
-                    c = c * k * scale;
-                    m = m * k * scale;
-                    y = y * k * scale;
+                    k *= scale;
+                    c *= k;
+                    m *= k;
+                    y *= k;
                 }
             }