--- a/src/ImageSharp/Formats/Jpeg/Components/Decoder/ColorConverters/JpegColorConverter.FromGrayScaleVector.cs
+++ b/src/ImageSharp/Formats/Jpeg/Components/Decoder/ColorConverters/JpegColorConverter.FromGrayScaleVector.cs
@@ -9,9 +9,9 @@ namespace SixLabors.ImageSharp.Formats.Jpeg.Components.Decoder.ColorConverters
 {
     internal abstract partial class JpegColorConverterBase
     {
-        internal sealed class FromGrayScaleVector8 : VectorizedJpegColorConverter
+        internal sealed class FromGrayScaleVector : VectorizedJpegColorConverter
         {
-            public FromGrayScaleVector8(int precision)
+            public FromGrayScaleVector(int precision)
                 : base(JpegColorSpace.Grayscale, precision)
             {
             }
@@ -23,7 +23,7 @@ namespace SixLabors.ImageSharp.Formats.Jpeg.Components.Decoder.ColorConverters
 
                 var scale = new Vector<float>(1 / this.MaximumValue);
 
-                nint n = values.Component0.Length / 8;
+                nint n = values.Component0.Length / Vector<float>.Count;
                 for (nint i = 0; i < n; i++)
                 {
                     ref Vector<float> c0 = ref Unsafe.Add(ref cBase, i);