### Prerequisites

- [x] I have written a descriptive pull-request title
- [x] I have verified that there are no overlapping [pull-requests](https://github.com/SixLabors/ImageSharp/pulls) open
- [x] I have verified that I am following the existing coding patterns and practice as demonstrated in the repository. These follow strict Stylecop rules :cop:.
- [x] I have provided test coverage for my change (where applicable)

### Description
Automatic AVX/SSE implementation for jpeg color converters based on `Vector` API. This PR uses existing code but with variable vector size rather than explicit `8`.

Master
|                              Method |                Job |       Mean |     Error |    StdDev | Ratio |
|------------------------------------ |------------------- |-----------:|----------:|----------:|------:|
|        'Baseline 4:4:4 Interleaved' | 1. No HwIntrinsics | 161.405 ms | 1.0435 ms | 0.8714 ms |  1.00 |
|        'Baseline 4:4:4 Interleaved' |             2. SSE |  29.430 ms | 0.1136 ms | 0.0949 ms |  0.18 |
|        'Baseline 4:4:4 Interleaved' |             3. AVX |  11.229 ms | 0.0369 ms | 0.0346 ms |  0.07 |
|                                     |                    |            |           |           |       |
|        'Baseline 4:2:0 Interleaved' | 1. No HwIntrinsics |  69.386 ms | 0.4476 ms | 0.4187 ms |  1.00 |
|        'Baseline 4:2:0 Interleaved' |             2. SSE |  14.439 ms | 0.0582 ms | 0.0516 ms |  0.21 |
|        'Baseline 4:2:0 Interleaved' |             3. AVX |   8.542 ms | 0.0390 ms | 0.0365 ms |  0.12 |
|                                     |                    |            |           |           |       |
|        'Baseline 4:0:0 (grayscale)' | 1. No HwIntrinsics |  12.338 ms | 0.2449 ms | 0.2405 ms |  1.00 |
|        'Baseline 4:0:0 (grayscale)' |             2. SSE |   3.307 ms | 0.0115 ms | 0.0108 ms |  0.27 |
|        'Baseline 4:0:0 (grayscale)' |             3. AVX |   1.566 ms | 0.0063 ms | 0.0056 ms |  0.13 |
|                                     |                    |            |           |           |       |
| 'Progressive 4:2:0 Non-Interleaved' | 1. No HwIntrinsics | 118.681 ms | 0.5633 ms | 0.5269 ms |  1.00 |
| 'Progressive 4:2:0 Non-Interleaved' |             2. SSE |  24.615 ms | 0.1012 ms | 0.0947 ms |  0.21 |
| 'Progressive 4:2:0 Non-Interleaved' |             3. AVX |  13.206 ms | 0.0487 ms | 0.0407 ms |  0.11 |

PR
|                              Method |                Job |       Mean |     Error |    StdDev | Ratio |
|------------------------------------ |------------------- |-----------:|----------:|----------:|------:|
|        'Baseline 4:4:4 Interleaved' | 1. No HwIntrinsics | 219.807 ms | 1.1166 ms | 0.9898 ms |  1.00 |
|        'Baseline 4:4:4 Interleaved' |             2. SSE |  28.429 ms | 0.1111 ms | 0.0985 ms |  0.13 |
|        'Baseline 4:4:4 Interleaved' |             3. AVX |  11.344 ms | 0.0534 ms | 0.0474 ms |  0.05 |
|                                     |                    |            |           |           |       |
|        'Baseline 4:2:0 Interleaved' | 1. No HwIntrinsics |  91.371 ms | 0.1973 ms | 0.1749 ms |  1.00 |
|        'Baseline 4:2:0 Interleaved' |             2. SSE |  13.635 ms | 0.0617 ms | 0.0577 ms |  0.15 |
|        'Baseline 4:2:0 Interleaved' |             3. AVX |   8.729 ms | 0.1201 ms | 0.1065 ms |  0.10 |
|                                     |                    |            |           |           |       |
|        'Baseline 4:0:0 (grayscale)' | 1. No HwIntrinsics |  12.822 ms | 0.0957 ms | 0.0895 ms |  1.00 |
|        'Baseline 4:0:0 (grayscale)' |             2. SSE |   3.299 ms | 0.0114 ms | 0.0107 ms |  0.26 |
|        'Baseline 4:0:0 (grayscale)' |             3. AVX |   1.551 ms | 0.0086 ms | 0.0080 ms |  0.12 |
|                                     |                    |            |           |           |       |
| 'Progressive 4:2:0 Non-Interleaved' | 1. No HwIntrinsics | 176.093 ms | 0.4813 ms | 0.4502 ms |  1.00 |
| 'Progressive 4:2:0 Non-Interleaved' |             2. SSE |  22.979 ms | 0.0551 ms | 0.0430 ms |  0.13 |
| 'Progressive 4:2:0 Non-Interleaved' |             3. AVX |  13.356 ms | 0.0924 ms | 0.0864 ms |  0.08 |

SSE speedup might not be that big but it's a nice 1ms saving for each case.

Don't look at scalar implementation performance 'downgrade' - BenchmarkDotNet decided to do a 1 iteration run for those so it's irrelevant.

----------------------------------------------------------------

PR also includes tests and **very little** performance tweaks - I've removed some if checks from converters for simd size as it's guaranteed that JPEG images have a width divisible by 8 without a remainder as it consists of 8x8 blocks. Once again, speedup is not that big but it's always nice to be a little bit faster with less code.

Master
|                              Method |      Mean |     Error |    StdDev |
|------------------------------------ |----------:|----------:|----------:|
|        'Baseline 4:4:4 Interleaved' | 11.504 ms | 0.0719 ms | 0.0673 ms |
|        'Baseline 4:2:0 Interleaved' |  8.605 ms | 0.0372 ms | 0.0348 ms |
|        'Baseline 4:0:0 (grayscale)' |  1.575 ms | 0.0054 ms | 0.0042 ms |
| 'Progressive 4:2:0 Non-Interleaved' | 13.207 ms | 0.0472 ms | 0.0369 ms |

PR
|                              Method |      Mean |     Error |    StdDev |
|------------------------------------ |----------:|----------:|----------:|
|        'Baseline 4:4:4 Interleaved' | 11.545 ms | 0.0444 ms | 0.0394 ms |
|        'Baseline 4:2:0 Interleaved' |  8.451 ms | 0.0450 ms | 0.0421 ms |
|        'Baseline 4:0:0 (grayscale)' |  1.519 ms | 0.0017 ms | 0.0015 ms |
| 'Progressive 4:2:0 Non-Interleaved' | 13.031 ms | 0.0201 ms | 0.0168 ms |

For whatever reason first image benchmark is **very** unstable, it goes from 11.0ms up to 11.5ms on master and this PR. Other images have a stable speedup (at least on simple benchmark config, multi-intrinsic one is a bit unstable for all test cases). Maybe it's caused by the GC. Baseline 444 image allocates 6 times more bytes than Baseline 420 and Baseline 400.