@brianpopow 2021-12-02
> Don't look at scalar implementation performance 'downgrade' - BenchmarkDotNet decided to do a 1 iteration run for those so it's irrelevant.

Maybe its worth changing the benchmark Config from `.WithLaunchCount(1)` to `.WithLaunchCount(3)` to get more stable results? It will take longer of course.