@br3aker 2021-12-02
> Maybe its worth changing the benchmark Config from .WithLaunchCount(1) to .WithLaunchCount(3) to get more stable results? It will take longer of course.

Scalar code wasn't touched by this PR as it's purely about automatic SSE converters and AVX converters optimization. So it's actually irrelevant but we have a scalar/sse/avx config so I've decided  to leave it as is. AFAIK BenchmarkDotNet don't want to benchmark code with such high magnitude (and we don't need it either so....)