@br3aker 2021-12-02
It's actually phenomenal how much performance was gained via intrinsics!