@JimBobSquarePants 2021-12-02
@br3aker Looks like we've got some failing tests in NET472

https://github.com/SixLabors/ImageSharp/runs/4393171045?check_suite_focus=true#step:12:78