@br3aker 2021-12-03
This is very strange, I've done nothing to the actual convertion code yet it have passed net472 tests.