@br3aker 2021-12-03
@JimBobSquarePants okay, I'm almost 100% sure there's some sort of a bug here. I can reproduce failing tests locally only on `core 2.1`. `net472`, `core 3.1` and `net5` are all passing. Furthermore, only remote executor tests are failing, `AllowAll` executes in the current process without remote execution and it passes for `Vector<T>` implementations, `EnableAVX 0` for `Vector<float>.Count == 4` aka auto SSE fails with some absolutely random values, it just can't happen - it's the same code with `Vector<float>`. Especially considering other passing runtimes.

What's our next step on this?