@br3aker 2021-12-03
Yet another detail: remote fails on 4th index - start of the second SSE vector.