@br3aker 2021-12-03
@brianpopow I'm either dumb or don't understand how remote executor works. Code below **fails** on `core2.1` and `net472`, works on every other runtime:
![image](https://user-images.githubusercontent.com/20967409/144630503-b844d1de-8bef-42b4-89eb-6d69edfb7274.png)


```
[Fact]
public void FoolproofFact()
{
    FeatureTestRunner.RunWithHwIntrinsicsFeature(
        RunTest,
        HwIntrinsics.DisableAVX);

    static void RunTest()
    {
        Assert.Equal(4, Vector<float>.Count);
    }
}
```

This test **does NOT fail** on any of the runtimes:
```
[Fact]
public void FoolproofFact()
{
    FeatureTestRunner.RunWithHwIntrinsicsFeature(
        RunTest,
        HwIntrinsics.DisableSIMD);

    static void RunTest()
    {
        Assert.Equal(4, Vector<float>.Count);
    }
}
```

Looks like 4 is a minimum hardware vector size. Wondering if JIT could compile Vector<float> into SSE instructions with `Vector<float>.Count == 4` - that would explain index 4 error.