@br3aker 2021-12-04
Another detail related to this problem, some code in repo uses `Vector<T>.Count` API in conjunction with the x86 intrinsics - they both are under `SUPPORTS_RUNTIME_INTRINSICS` conditional. Example: [png filters](https://github.com/SixLabors/ImageSharp/tree/master/src/ImageSharp/Formats/Png/Filters). That conditional is set for core 3.0 and up - platforms which do not suffer from this strange error I've encountered. That's why tests with existing code pass - they do not run remote executor for `Vector<T>.Count` on net472 and core20.

Marking this PR WIP until we decide how we should approach this.