@brianpopow 2021-12-06
@br3aker I think the reason for the failing tests is, that they are not inside a `#if SUPPORTS_RUNTIME_INTRINSICS` block, e.g. `FromYccKVector`. The test runs with and without AVX.