@br3aker 2021-12-06
@brianpopow `Vector<T>` implementation for 8 floats wasn't excluded from the net472/core21 via `#if` before this PR so it will be a regression. Which is actually strange, a lot of simd utils code with `Vector<T>` is covered by `#if`. It clearly works without remote executor at least on my machine and it did work on CI.

Honestly, if it was my call - I'd say it's good to go without tests with disabled avx for SSE check. If it works on other runtimes without a problem and work with avx hardware - sse should not be a problem, at least from out side.