@br3aker 2021-12-06
> If we somehow manage to resolve the challenge of Vector<float>.Count without loosing coverage I'm fine with it though.

It doesn't affect coverage afaik. The only problem atm is that automatic SSE implementation by jit from `Vector<float>` code won't be tested if CI hardware supports avx. I actually not even sure that there's any CPUs right now which don't support at least minimal AVX set.