@JimBobSquarePants 2021-12-12
@brianpopow weird... That test I thought we just disabled failed here. https://github.com/SixLabors/ImageSharp/runs/4497782950?check_suite_focus=true#step:13:277