@brianpopow 2021-12-12
> @brianpopow weird... That test I thought we just disabled failed here. https://github.com/SixLabors/ImageSharp/runs/4497782950?check_suite_focus=true#step:13:277

I dont understand how this test gets executed `AllocateSingleMemoryOwner_Finalization_ReturnsToPool` on macOS.

Also it seems there are other tests failing, not just this one: `AllocateMemoryGroup_Finalization_ReturnsToPool`